backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Ze Backend API"
authors = ["Ze <ze@ze.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111"
uvicorn = "^0.30"
python-dotenv = "^1.1.1"
twelvedata = "^1.2.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
